feat!: Use the new API project with breaking changes
