+++ package.json
@@ -13,7 +13,6 @@
   "author": "Lenra",
   "license": "MIT",
   "dependencies": {
-    "@lenra/app-server": "^1.0.0-beta.1",
-    "@lenra/components": "^1.0.0-beta.5"
+    "@lenra/app": "^1.0.0-beta.12"
   }
 }
