--- src/index.gen.js
@@ -1,15 +1,0 @@
-export const views = {
-	"counter": "counter",
-	"home": "home",
-	"json": {
-		"counter": "json.counter"
-	},
-	"main": "main",
-	"menu": "menu"
-};
-export const listeners = {
-	"increment": "increment",
-	"onEnvStart": "onEnvStart",
-	"onSessionStart": "onSessionStart",
-	"onUserFirstJoin": "onUserFirstJoin"
-};
