+++ src/views/counter.js
@@ -1,22 +1,18 @@
-import { Flex, Text, Button } from "@lenra/components";
-import { listeners } from "../index.gen.js";
+import { Listener } from "@lenra/app";
+import { Counter } from "../classes/Counter.js";
 
 /**
  * 
- * @param {import("../classes/Counter").Counter[]} param0 
- * @param {import("@lenra/app-server").props} param1 
- * @returns 
+ * @param {Counter[]} param0 
+ * @param {*} _props 
+ * @returns {import("@lenra/app").JsonViewResponse}
  */
-export default function ([counter], { text }) {
-  return Flex([
-    Text(`${text}: ${counter.count}`),
-    Button("+")
-      .onPressed(listeners.increment, {
-        "id": counter._id
+export default function ([counter], _props) {
+  return {
+    value: counter.count,
+    onIncrement: Listener("increment")
+      .props({
+        id: counter._id
       })
-  ])
-    .spacing(16)
-    .mainAxisAlignment("spaceEvenly")
-    .crossAxisAlignment("center")
+  };
 }
-
