+++ src/views/lenra/main.js
@@ -1,10 +1,9 @@
-import { Flex, View } from "@lenra/components";
-import { views } from "../index.gen.js";
+import { Flex, View } from "@lenra/app";
 
 export default function (_data, _props) {
   return Flex([
-    View(views.menu),
-    View(views.home)
+    View("lenra.menu"),
+    View("lenra.home")
   ])
     .direction("vertical")
     .scroll(true)
